Promenjeni longovi u integer

+++ uber-clone/src/main/java/com/backend/uberclone/model/Review.java
@@ -6,7 +6,6 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
-import java.io.CharArrayReader;
 
 @Getter
 @Setter
@@ -17,7 +16,7 @@ public class Review {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name="ride_id")
